pyproject.toml
[tool.poetry]
name = "graphite-image-ticket-to-calendar"
version = "0.1.0"
description = "Image processor with Graphite"
authors = ["Yuriy Mud <yuriy.mud3@gmail.com>"]

license = "Mozilla Public License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
grafi = "^0.0.8"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
requests = "^2.32.3"
onnxruntime = "1.19.2"
google-api-python-client = "^2.166.0"
google-auth = "^2.38.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.2"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

